0911 supervisor complete: show list;get/set quiz and question

+++ backend/src/main/java/berryStreet/bluering/backend/entity/Option.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class Option {
-    private int key;
+    private long key;
     private String value;
     private int point;
 }
